fix: issues after migration
